requirements.txt
# BE FORWARD Web Scraper - Python Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.1.0
tqdm>=4.66.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
